backend/workmatch/requirements.txt
# Core ADK framework
google-adk

google-cloud-secret-manager
google-cloud-storage
google-cloud-aiplatform

# Pinecone for semantic search
pinecone

# Web framework (included in ADK Web)
fastapi
uvicorn

# Telemetry

langfuse
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-api